+++ website/gatsby-node.js
@@ -1,41 +1,37 @@
-// NOTE: It is possible to override the ocular-provided callbacks
-// and this take control any aspect of gatsby:
+const resolve = require('path').resolve;
 
-// exports.onCreateNode = ({ node, actions, getNode }) =>
-//   ocular.onCreateNode({ node, actions, getNode });
+const DEPENDENCIES = require('./package.json').dependencies;
+const ALIASES = require('ocular-dev-tools/config/ocular.config')({
+  root: resolve(__dirname, '..')
+}).aliases;
 
-// exports.setFieldsOnGraphQLNodeType = ({ type, actions }) =>
-//   ocular.setFieldsOnGraphQLNodeType({ type, actions });
+// When duplicating example dependencies in website, autogenerate
+// aliases to ensure the website version is picked up
+// NOTE: module dependencies are automatically injected
+// TODO - should this be automatically done by ocular-gatsby?
+const dependencyAliases = {};
+for (const dependency in DEPENDENCIES) {
+  dependencyAliases[dependency] = `${__dirname}/node_modules/${dependency}`;
+}
 
-// // This is a main gatsby entry point
-// // Here we get to programmatically create pages after all nodes are created
-// // by gatsby.
-// // We use graphgl to query for nodes and iterate
-// exports.createPages = ({ graphql, actions }) =>
-//   ocular.createPages({ graphql, actions });
-const ocularConfig = require('./ocular-config');
-const getGatsbyNodeCallbacks = require('ocular-gatsby/gatsby-node');
-
-const callbacks = getGatsbyNodeCallbacks(ocularConfig);
-
-module.exports = callbacks;
-
-const onCreateWebpackConfig = callbacks.onCreateWebpackConfig;
-
-callbacks.onCreateWebpackConfig = function onCreateWebpackConfigOverride(opts) {
-  onCreateWebpackConfig(opts);
-
+module.exports.onCreateWebpackConfig = function onCreateWebpackConfigOverride(opts) {
   const {
-    // stage, // build stage: ‘develop’, ‘develop-html’, ‘build-javascript’, or ‘build-html’
+    stage, // build stage: ‘develop’, ‘develop-html’, ‘build-javascript’, or ‘build-html’
     // rules, // Object (map): set of preconfigured webpack config rules
     // plugins, // Object (map): A set of preconfigured webpack config plugins
-    // getConfig, // Function that returns the current webpack config
+    getConfig, // Function that returns the current webpack config
     loaders, // Object (map): set of preconfigured webpack config loaders
     actions
   } = opts;
 
   console.log(`App rewriting gatsby webpack config`); // eslint-disable-line
 
+  const config = getConfig();
+  config.resolve = config.resolve || {};  
+  config.resolve.alias = config.resolve.alias || {};  
+  Object.assign(config.resolve.alias, ALIASES, dependencyAliases);
+
+  /*
   // Recreate it with custom exclude filter
   // Called without any arguments, `loaders.js` will return an
   // object like:
@@ -56,24 +52,32 @@ callbacks.onCreateWebpackConfig = function onCreateWebpackConfigOverride(opts) {
     // Exclude all node_modules from transpilation, except for ocular
     exclude: modulePath =>
       /node_modules/.test(modulePath) &&
-      !/node_modules\/(ocular|ocular-gatsby|gatsby-plugin-ocular)/.test(modulePath)
+      !/node_modules\/(ocular|ocular-gatsby|gatsby-theme-ocular)/.test(modulePath)
   });
+
+  // Omit the default rule where test === '\.jsx?$'
+  const rules = [newJSRule];
+
+  if (stage === 'build-html') {
+    rules.push({
+      test: /mapbox-gl/,
+      use: loaders.null()
+    });
+  }
 
   const newConfig = {
     module: {
-      rules: [
-        // Omit the default rule where test === '\.jsx?$'
-        newJSRule
-      ]
+      rules
     },
     node: {
       fs: 'empty'
-    },
-    resolve: {
-      alias: Object.assign(ocularConfig.webpack.resolve.alias)
     }
   };
+  */
 
-  // Merges into the webpack config
-  actions.setWebpackConfig(newConfig);
+  // Completely replace the webpack config for the current stage.
+  // This can be dangerous and break Gatsby if certain configuration options are changed.
+  // Generally only useful for cases where you need to handle config merging logic yourself,
+  // in which case consider using webpack-merge.
+  actions.setWebpackConfig(config);
 };
