--- website/gatsby-ssr.js
@@ -1,1 +1,0 @@
-module.exports.wrapPageElement = require('ocular-gatsby/gatsby-ssr').wrapPageElement;
